--- src/classes/TaskStatus.ts
@@ -1,6 +1,0 @@
-export const TaskStatus = {
-    Complete: "COMPLETE",
-    Incomplete: "INCOMPLETE",
-} as const;
-
-export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];
