+++ src/models/taskCheckOff.ts
@@ -1,7 +1,6 @@
 import { Model, DataTypes } from "sequelize";
 import type {
     CreationOptional,
-    Includeable,
     InferAttributes,
     InferCreationAttributes,
 } from "sequelize";
@@ -12,11 +11,6 @@ import { ScheduleDatabase } from "../classes/scheduleDatabase.ts";
 import { Task } from "./task.ts";
 import { Employee } from "./employee.ts";
 import { User } from "./user.ts";
-
-const include: Includeable = {
-    model: Employee,
-    include: [User],
-};
 
 export class TaskCheckOff extends Model<
     InferAttributes<TaskCheckOff>,
@@ -80,7 +74,10 @@ class TaskCheckOffRouter extends ModelRouter {
                 req,
                 res,
                 {
-                    include,
+                    include: {
+                        model: Employee,
+                        include: [User],
+                    },
                 },
                 "id",
             ),
@@ -91,7 +88,12 @@ class TaskCheckOffRouter extends ModelRouter {
                 req,
                 res,
                 {
-                    include,
+                    include: [
+                        {
+                            model: Employee,
+                            include: [User],
+                        },
+                    ],
                 },
                 "taskID",
             ),
