+++ src/routers/taskRouter.ts
@@ -1,0 +1,79 @@
+import { Router } from "express";
+import { ModelRouter } from "../classes/databaseModel.ts";
+import { ScheduleDatabase } from "../classes/scheduleDatabase.ts";
+import { Task } from "../models/task.ts";
+import { TaskCheckOff } from "../models/taskCheckOff.ts";
+
+// i've discovered that we need to move routers out of model files for certain things to work...
+class TaskRouter extends ModelRouter {
+    public path(): string {
+        return "/tasks";
+    }
+
+    protected buildRouter(router: Router): void {
+        router.post("/", (req, res) => ScheduleDatabase.create(Task, req, res));
+        router.put("/", (req, res) =>
+            ScheduleDatabase.update(Task, req, res, "id"),
+        );
+        router.delete("/:id", (req, res) =>
+            ScheduleDatabase.delete(Task, req, res, "id"),
+        );
+        router.get("/:id", (req, res) =>
+            ScheduleDatabase.getWhere(
+                Task,
+                req,
+                res,
+                { include: TaskCheckOff },
+                "id",
+            ),
+        );
+        router.get("/taskList/:taskListID", (req, res) =>
+            ScheduleDatabase.getAllWhere(
+                Task,
+                req,
+                res,
+                { include: TaskCheckOff },
+                "taskListID",
+            ),
+        );
+    }
+
+    public async createOrUpdateTask(
+        task: Task,
+        listOrder: number,
+    ): Promise<void> {
+        let id = task.id;
+        task.listOrder = listOrder;
+        let taskPromises: Promise<any>[] = [];
+
+        console.log("creating task: " + JSON.stringify(task));
+        if (task.id > 0)
+            taskPromises.push(Task.update(task, { where: { id } }));
+        else {
+            let taskInstance = await Task.create(task);
+            id = taskInstance.id;
+        } 
+        
+        console.log("task created");
+
+        const checkOffs: TaskCheckOff[] = (task as any)["checkOffs"];
+
+        if (checkOffs) {
+            checkOffs.forEach((check) => {
+                check.taskID = id;
+
+                if (check.id > 0)
+                    taskPromises.push(
+                        TaskCheckOff.update(check, {
+                            where: { id: check.id },
+                        }),
+                    );
+                else taskPromises.push(TaskCheckOff.create(check));
+            });
+        }
+
+        await Promise.all(taskPromises);
+    }
+}
+
+export const taskRouter = new TaskRouter();
